dev_requirements.txt
# Deis controller requirements
azure==0.7.0
boto==2.11.0
celery==3.0.22  # 3.0.23 errors, see https://github.com/opdemand/deis/issues/148
Django==1.5.2
django-allauth==0.13.0
django-celery==3.0.23
django-json-field==0.5.5
django-yamlfield==0.4
djangorestframework==2.3.7
gevent==0.13.8
gunicorn==18.0
paramiko==1.11.0
psycopg2==2.5.1
pycrypto==2.6
PyYAML==3.10
South==0.8.2

# Deis client requirements
docopt==0.6.1
#PyYAML==3.10
requests==1.2.3

# Deis documentation requirements
Sphinx>=1.2b1
smartypants>=1.8.2
sphinxcontrib-httpdomain>=1.1.9

# Run "make coverage" for the % of code exercised during tests
coverage>=3.6

# Generates template documentation in the Django admin
docutils>=0.11

# Run "make flake8" to check python syntax and style
flake8>=2.0

# Python expect is used to automate client-driven tests
pexpect-u>=2.5.1
